+++ jackplay/src/jackplay/bootstrap/Trace.java
@@ -2,12 +2,12 @@ package jackplay.bootstrap;
 
 import java.util.Date;
 
-class TraceLog {
+class Trace {
     Date when;
     long whenAsTimeMs;
 
     // mandatory for all entries
-    TracePoint tracePoint;
+    Site site;
     PlayGround pg;
     long threadId;
     String threadName;
@@ -25,13 +25,19 @@ class TraceLog {
 
     String exceptionStackTrace;      // exception only
 
-    TraceLog(TracePoint tracePoint, PlayGround pg, String uuid) {
+    Trace(Site site, PlayGround pg, String uuid) {
         this.when = new Date();
         this.whenAsTimeMs = System.currentTimeMillis();
-        this.tracePoint = tracePoint;
+        this.site = site;
         this.pg = pg;
         this.threadId = Thread.currentThread().getId();
         this.threadName = Thread.currentThread().getName();
         this.uuid = uuid;
     }
+}
+
+enum Site {
+    MethodEntry,
+    MethodReturns,
+    MethodThrowsException
 }