+++ jackplay/src/jackplay/bootstrap/TracePoint.java
@@ -1,7 +1,0 @@
-package jackplay.bootstrap;
-
-public enum TracePoint {
-    MethodEntry,
-    MethodReturns,
-    MethodThrowsException
-}